requirements.txt
# --- Core Traitement d'Images ---
torch>=2.0.1
torchvision>=0.15.2

# --- Interface Web ---
streamlit>=1.28.0

# --- Optimisation d'Hyperparamètres ---
optuna>=3.4.0

# --- Configuration Management ---
# PyYAML est le vrai nom du package pour utiliser 'import yaml'
PyYAML>=6.0.1

# --- Data Processing ---
numpy>=1.24.3
pandas>=2.0.3

# --- Image & Video Processing ---
opencv-python>=4.8.0.76
Pillow>=10.0.1

# --- Metrics & Visualization ---
matplotlib>=3.7.2
seaborn>=0.12.2
scikit-learn>=1.3.0

# --- Interpretability (LIME) ---
lime>=0.2.0
scikit-image>=0.20.0

# --- Audio (speech emotion + extraction) ---
transformers>=4.30.0
librosa>=0.10.0
soundfile>=0.12.0
imageio-ffmpeg>=0.4.9

# --- Utilities ---
tqdm>=4.66.1
